pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "marketplace_system"
version = "0.1.0"
description = "ETL system for marketplace API"
dependencies = [
    "apache-airflow>=2.11.0,<3.0.0",
    "requests",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-mock", "requests-mock"]

[tool.setuptools.packages]
find = { namespaces = false }

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"